use delete in range to clean overlap range

+++ src/raftstore/store/worker/region.rs
@@ -261,16 +261,24 @@ impl SnapContext {
         region
     }
 
-    fn handle_destroy(&mut self, region_id: u64, start_key: Vec<u8>, end_key: Vec<u8>) {
-        if let Err(e) = util::delete_all_files_in_range(&self.kv_db, &start_key, &end_key) {
-            error!(
-                "[region {}] failed to delete files in [{}, {}): {:?}",
-                region_id,
-                escape(&start_key),
-                escape(&end_key),
-                e
-            );
-            return;
+    fn handle_destroy(
+        &mut self,
+        region_id: u64,
+        start_key: Vec<u8>,
+        end_key: Vec<u8>,
+        use_delete_files: bool,
+    ) {
+        if use_delete_files {
+            if let Err(e) = util::delete_all_files_in_range(&self.kv_db, &start_key, &end_key) {
+                error!(
+                    "[region {}] failed to delete files in [{}, {}): {:?}",
+                    region_id,
+                    escape(&start_key),
+                    escape(&end_key),
+                    e
+                );
+                return;
+            }
         }
         if let Err(e) =
             util::delete_all_in_range(&self.kv_db, &start_key, &end_key, self.use_delete_range)
@@ -328,13 +336,20 @@ impl PendingDeleteRanges {
         ranges
     }
 
-    // filter out the overlap in ranges, after this function the ranges will not overlap
-    pub fn remove(&mut self, start_key: &[u8], end_key: &[u8]) {
+    // find any overlap range in [start_key, end_key),
+    // then trim the overlap ranges and return the non-overlap part in overlap ranges
+    pub fn trim_overlap_ranges(
+        &mut self,
+        start_key: &[u8],
+        end_key: &[u8],
+    ) -> Vec<(u64, Vec<u8>, Vec<u8>)> {
         if start_key >= end_key {
-            return;
+            return Vec::new();
         }
+        let mut to_destroy = Vec::new();
         let overlapped_ranges = self.find_overlapped_ranges(start_key, end_key);
         for (ref s_key, ref e_key) in overlapped_ranges {
+            let info = self.ranges.remove(s_key).unwrap();
             if s_key < &start_key.to_vec() {
                 if e_key <= &end_key.to_vec() {
                     // the right part of the range is trimed
@@ -343,8 +358,7 @@ impl PendingDeleteRanges {
                     //            [_ _ _ _ _ _ _ _ _ _)
                     //       start_key            end_key
                     //   [_ _ _ _ )                          after
-                    let info = self.ranges.get_mut(s_key).unwrap();
-                    info.end_key = start_key.to_vec();
+                    to_destroy.push((info.region_id, s_key.clone(), start_key.to_vec()));
                 } else {
                     // the middle part of the range is trimed
                     // s_key                   e_key
@@ -352,19 +366,16 @@ impl PendingDeleteRanges {
                     //           [_ _ _ _)
                     //      start_key   end_key
                     //   [_ _ _ _)       [_ _ _ _)        after
-                    let mut info = self.ranges.remove(s_key).unwrap();
-                    self.ranges.insert(end_key.to_vec(), info.clone());
-                    info.end_key = start_key.to_vec();
-                    assert!(self.ranges.insert(s_key.clone(), info).is_none());
+                    to_destroy.push((info.region_id, s_key.clone(), start_key.to_vec()));
+                    to_destroy.push((info.region_id, end_key.to_vec(), info.end_key.clone()));
                 }
             } else if e_key <= &end_key.to_vec() {
-                // the whole range is discarded
+                // the whole range is discarded, so do nothing
                 //     s_key    e_key
                 //       [_ _ _ _)             before
                 // start_key     end_key
                 //    [_ _ _ _ _ _ _)
                 //                             after
-                self.ranges.remove(s_key).unwrap();
             } else {
                 // the left part of the range is trimed
                 //         s_key         e_key
@@ -372,12 +383,14 @@ impl PendingDeleteRanges {
                 // start_key      end_key
                 //    [_ _ _ _ _ _ _)
                 //                  [_ _ _ )      after
-                let info = self.ranges.remove(s_key).unwrap();
-                assert!(self.ranges.insert(end_key.to_vec(), info).is_none());
+                to_destroy.push((info.region_id, end_key.to_vec(), info.end_key.clone()));
             }
         }
+
+        to_destroy
     }
 
+    // before an insert is called, must call trim_overlap_ranges to clean the overlap range
     pub fn insert(
         &mut self,
         region_id: u64,
@@ -385,7 +398,6 @@ impl PendingDeleteRanges {
         end_key: Vec<u8>,
         timeout: time::Instant,
     ) {
-        self.remove(&start_key, &end_key);
         let info = StalePeerInfo {
             region_id: region_id,
             end_key: end_key,
@@ -443,6 +455,15 @@ impl Runner {
             pending_delete_ranges: PendingDeleteRanges::default(),
         }
     }
+
+    pub fn destroy_overlap_ranges(&mut self, start_key: &[u8], end_key: &[u8]) {
+        let to_destroy = self.pending_delete_ranges
+            .trim_overlap_ranges(start_key, end_key);
+        for (region_id, s_key, e_key) in to_destroy {
+            self.ctx
+                .handle_destroy(region_id, s_key, e_key, false /* use_delete_files */);
+        }
+    }
 }
 
 impl Runnable<Task> for Runner {
@@ -462,7 +483,7 @@ impl Runnable<Task> for Runner {
                 if let Some(region) = self.ctx.handle_apply(region_id, status) {
                     let start_key = keys::enc_start_key(&region);
                     let end_key = keys::enc_end_key(&region);
-                    self.pending_delete_ranges.remove(&start_key, &end_key);
+                    self.destroy_overlap_ranges(&start_key, &end_key);
                 }
             }
             Task::Destroy {
@@ -479,10 +500,16 @@ impl Runnable<Task> for Runner {
                     );
                     // delay the range deletion because there might be a coprocessor request related to this range
                     let timeout = time::Instant::now() + self.clean_stale_peer_delay;
+                    self.destroy_overlap_ranges(&start_key, &end_key);
                     self.pending_delete_ranges
                         .insert(region_id, start_key, end_key, timeout);
                 } else {
-                    self.ctx.handle_destroy(region_id, start_key, end_key);
+                    self.ctx.handle_destroy(
+                        region_id,
+                        start_key,
+                        end_key,
+                        true, /* use_delete_files */
+                    );
                 }
             }
             Task::CleanStalePeer {} => {
@@ -491,7 +518,12 @@ impl Runnable<Task> for Runner {
                 let now = time::Instant::now();
                 let mut timeout_ranges = self.pending_delete_ranges.drain_timeout_ranges(now);
                 for (region_id, start_key, end_key) in timeout_ranges.drain(..) {
-                    self.ctx.handle_destroy(region_id, start_key, end_key);
+                    self.ctx.handle_destroy(
+                        region_id,
+                        start_key,
+                        end_key,
+                        true, /* use_delete_files */
+                    );
                 }
             }
         }
@@ -516,9 +548,7 @@ mod test {
 
     #[test]
     fn test_pending_delete_ranges_case1() {
-        let mut pending_delete_ranges = PendingDeleteRanges {
-            ranges: BTreeMap::new(),
-        };
+        let mut pending_delete_ranges = PendingDeleteRanges::default();
         let delay = Duration::from_millis(100);
         let id = 0;
 
@@ -526,7 +556,7 @@ mod test {
         pending_delete_ranges.insert(id, b"aa".to_vec(), b"bb".to_vec(), timeout);
         pending_delete_ranges.insert(id + 1, b"bb".to_vec(), b"cc".to_vec(), timeout);
 
-        // not reaching the timeout, so no regions
+        // not reaching the timeout, so no ranges
         let now = time::Instant::now();
         let ranges = pending_delete_ranges.drain_timeout_ranges(now);
         assert!(ranges.is_empty());
@@ -550,63 +580,7 @@ mod test {
 
     #[test]
     fn test_pending_delete_ranges_case2() {
-        let mut pending_delete_ranges = PendingDeleteRanges {
-            ranges: BTreeMap::new(),
-        };
-        let delay = Duration::from_millis(100);
-        let id = 0;
-
-        let timeout = time::Instant::now() + delay;
-        pending_delete_ranges.insert(id, b"aa".to_vec(), b"bb".to_vec(), timeout);
-
-        thread::sleep(delay / 2);
-
-        let timeout = time::Instant::now() + delay;
-        pending_delete_ranges.insert(id + 1, b"aa".to_vec(), b"bb".to_vec(), timeout);
-
-        thread::sleep(delay / 2);
-
-        // range has been updated, so will not timeout
-        let now = time::Instant::now();
-        let ranges = pending_delete_ranges.drain_timeout_ranges(now);
-        assert!(ranges.is_empty());
-
-        thread::sleep(delay / 2);
-
-        // now will timeout
-        let now = time::Instant::now();
-        let ranges = pending_delete_ranges.drain_timeout_ranges(now);
-        assert_eq!(ranges, [(id + 1, b"aa".to_vec(), b"bb".to_vec())]);
-    }
-
-    #[test]
-    fn test_pending_delete_ranges_case3() {
-        let mut pending_delete_ranges = PendingDeleteRanges {
-            ranges: BTreeMap::new(),
-        };
-        let delay = Duration::from_millis(100);
-        let id = 0;
-
-        let timeout = time::Instant::now() + delay;
-        pending_delete_ranges.insert(id, b"aa".to_vec(), b"bb".to_vec(), timeout);
-
-        thread::sleep(delay / 2);
-
-        pending_delete_ranges.remove(&b"aa".to_vec(), &b"bb".to_vec());
-
-        thread::sleep(delay / 2);
-
-        // range has been removed
-        let now = time::Instant::now();
-        let ranges = pending_delete_ranges.drain_timeout_ranges(now);
-        assert!(ranges.is_empty());
-    }
-
-    #[test]
-    fn test_pending_delete_ranges_case4() {
-        let mut pending_delete_ranges = PendingDeleteRanges {
-            ranges: BTreeMap::new(),
-        };
+        let mut pending_delete_ranges = PendingDeleteRanges::default();
         let delay = Duration::from_millis(100);
         let id = 0;
 
@@ -620,6 +594,14 @@ mod test {
         thread::sleep(delay / 2);
 
         let timeout = time::Instant::now() + delay;
+        let left = pending_delete_ranges.trim_overlap_ranges(&b"g".to_vec(), &b"q".to_vec());
+        assert_eq!(
+            left,
+            [
+                (id + 1, b"f".to_vec(), b"g".to_vec()),
+                (id + 1, b"q".to_vec(), b"t".to_vec()),
+            ]
+        );
         pending_delete_ranges.insert(id + 2, b"g".to_vec(), b"q".to_vec(), timeout);
 
         thread::sleep(delay / 2);
@@ -630,8 +612,6 @@ mod test {
             ranges,
             [
                 (id, b"a".to_vec(), b"c".to_vec()),
-                (id + 1, b"f".to_vec(), b"g".to_vec()),
-                (id + 1, b"q".to_vec(), b"t".to_vec()),
                 (id, b"x".to_vec(), b"z".to_vec()),
             ]
         );
@@ -644,7 +624,7 @@ mod test {
     }
 
     #[test]
-    fn test_pending_delete_ranges_case5() {
+    fn test_pending_delete_ranges_case3() {
         let mut pending_delete_ranges = PendingDeleteRanges {
             ranges: BTreeMap::new(),
         };
@@ -658,6 +638,15 @@ mod test {
 
         thread::sleep(delay / 2);
 
+        let left = pending_delete_ranges.trim_overlap_ranges(&b"g".to_vec(), &b"q".to_vec());
+        assert_eq!(
+            left,
+            [
+                (id + 1, b"f".to_vec(), b"g".to_vec()),
+                (id + 1, b"q".to_vec(), b"t".to_vec()),
+            ]
+        );
+
         let timeout = time::Instant::now() + delay;
         pending_delete_ranges.insert(id + 2, b"g".to_vec(), b"q".to_vec(), timeout);
 
@@ -669,8 +658,6 @@ mod test {
             ranges,
             [
                 (id, b"a".to_vec(), b"c".to_vec()),
-                (id + 1, b"f".to_vec(), b"g".to_vec()),
-                (id + 1, b"q".to_vec(), b"t".to_vec()),
                 (id, b"x".to_vec(), b"z".to_vec()),
             ]
         );
